requirements.txt
aiofiles==24.1.0
aiosqlite==0.20.0
discord.py==2.4.0
nest_asyncio==1.6.0
python-dotenv==1.0.1
